requirements.txt
google-generativeai 
phidata
pillow